Add activate function & email sending

+++ user_control/views.py
@@ -1,5 +1,7 @@
 import jwt
 import requests
+from django.http import HttpResponse
+
 from .models import JwtModel, CustomUser, Favorite
 from datetime import datetime, timedelta
 from django.conf import settings
@@ -16,6 +18,14 @@ from rest_framework.viewsets import ModelViewSet
 import re
 import json
 from django.db.models import Q, Count, Subquery, OuterRef
+
+from django.contrib.auth import get_user_model
+from django.utils.http import urlsafe_base64_encode, urlsafe_base64_decode
+from django.contrib.sites.shortcuts import get_current_site
+from .tokens import account_activation_token
+from django.core.mail import send_mail, EmailMessage
+from django.utils.encoding import force_bytes, force_str
+from django.template.loader import render_to_string
 
 
 def get_random(length):
@@ -64,6 +74,8 @@ class LoginView(APIView):
 
         if not user:
             return Response({"error": "Invalid username or password"}, status="400")
+        if not user.is_active:
+            return Response({"error": "Email is not verified"}, status="433")
 
         JwtModel.objects.filter(user_id=user.id).delete()
 
@@ -95,8 +107,45 @@ class RegisterView(APIView):
         serializer.is_valid(raise_exception=True)
 
         CustomUser.objects._create_user(**serializer.validated_data)
+        user = CustomUser.objects.get(email=serializer.validated_data["email"])
 
-        return Response({"success": "User created."}, status=201)
+        current_site = get_current_site(request)
+        mail_subject = 'Activate your account.'
+        message = render_to_string('email_template.html', {
+            #'user': user,
+            'domain': current_site.domain,
+            'uid': urlsafe_base64_encode(force_bytes(user.id)),
+            'token': account_activation_token.make_token(user),
+        })
+        to_email = serializer.validated_data["email"]
+        email = EmailMessage(
+            mail_subject,
+            message,
+            'takadykova@edu.hse.ru',
+            [to_email],
+        )
+
+        #email.fail_silently = True
+        email.send()
+
+        #send_mail(mail_subject, message, 'takadykova@edu.hse.ru', [to_email])
+        return Response({"success": "Please confirm your email address to complete the registration",
+                         "message": message,
+                         "to_email": to_email}, status=201)
+
+
+def activate(request, uidb64, token):
+    try:
+        uid = force_str(urlsafe_base64_decode(uidb64))
+        user = CustomUser.objects.get(id=uid)
+    except Exception as e:
+        user = None
+    if user is not None and account_activation_token.check_token(user, token):
+        user.is_active = True
+        user.save()
+        return HttpResponse('Thank you for your email confirmation. Now you can login your account.')
+    else:
+        return HttpResponse('Activation link is invalid!')
 
 
 class RefreshView(APIView):
